helper moved into aa2cg.py
